+++ CHANGELOG.md
@@ -8,7 +8,6 @@ y este proyecto utiliza [SemVer](https://semver.org/lang/es/).
 ## [Unreleased]
 
 ### Added
-- Added login feature for backoffice
 
 ### Changed
 
@@ -16,7 +15,23 @@ y este proyecto utiliza [SemVer](https://semver.org/lang/es/).
 
 ### Fixed
 
-## [v0.0.9] - 2025-04-06
+---
+
+## [0.1.2] - 2025-04-06
+
+### Added
+- Added logic to fetch program-panelist relations
+- Added endpoint toclean cache for specific panelist
+
+### Fixed
+- Fixed panelists bio, now it is optional
+
+## [0.1.1] - 2025-04-06
+
+### Added
+- Added login feature for backoffice
+
+## [0.1.0] - 2025-04-06
 
 ### Added
 - Added redis caché service for schedule GETs
@@ -24,12 +39,12 @@ y este proyecto utiliza [SemVer](https://semver.org/lang/es/).
 ### Removed
 - Removed changelog modification from release script
 
-## [v0.0.9] - 2025-04-06
+## [0.0.9] - 2025-04-06
 
 ### Fixed
 - Fixed cors and 502 errors in prod
 
-## [v0.0.8] - 2025-04-06
+## [0.0.8] - 2025-04-06
 
 ### Added
 - Added coverage check to PRs